requirements.txt
Django 3.2
django-phonenumber-field 5.1.0
Pillow-8.2.0
psycopg2-2.8.6
sorl_thumbnail-12.7.0






